Some obvious factoring
